Remove comments from code that probably should never have been commented out

+++ MyApplication/app/src/main/java/ledge/muscleup/presentation/MainActivity.java
@@ -185,7 +185,7 @@ public class MainActivity extends Activity {
 
             File outFile = new File(copyPath);
 
-            //if (!outFile.exists()) {
+            if (!outFile.exists()) {
                 InputStreamReader in = new InputStreamReader(assetManager.open(asset));
                 FileWriter out = new FileWriter(outFile);
 
@@ -197,7 +197,7 @@ public class MainActivity extends Activity {
 
                 out.close();
                 in.close();
-            //}
+            }
         }
     }
 }
